+++ audio_processing/source_seperation/asr_testing/google_cloud_key.json
@@ -1,0 +1,12 @@
+{
+  "type": "service_account",
+  "project_id": "focused-service-291115",
+  "private_key_id": "def9cae625ed30384b9edc529a4192ec7b42fcc6",
+  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDuGvkq2lvIDaHb\nzm3YB8VAC23l46qEuG6Npoiz1dzuLcmZJlwJspsW+Wous1QjtZO4KETwjVFRqIco\nYh/hiHmfCHGI8HSTS66JklXaiPZbBkojlmUcVozBN++ZQ7pdy+ynNVuAILbrjV5v\nve2lj3NOaOq8sKu8FGRTfEUHmEjg+0yFBnbwfLj9B9AKrUt8uwwV+X2+f+cmRzZv\nK90lmp6tLP6kEIJgCK/QA7y9T79gjzFBrp1IkanLZIz3jletLaEkmy7VXX2RUfz1\ntXZy/a6kW1NQd4V8G1sJQXKN4NDjSXkH2/A7GRCAQ+VGOL1D45djx3ISaT7F7Ega\nzDfUnYapAgMBAAECggEAYfQEx2VXVsCM8tFEjRf9RLCvAGEdmEYLbQiyulDj97Qs\neG24dWDdvqDS1u19UsdzQtKGMfPhHVfYanubo/uyFQOzOWO6gZIj9oKWATZRKvc2\nbteDNxpAA4kQb5vWQpoO/GbOArLyrYrV6+BVsMOH6xzsXL3NmK1G7LV17H7wlGTi\nSLLacd+V9qYbICjdawrzlFMH4gAMwxUYZHpmWwajp2P5OCEPu+Mz9GXOw0HoJr+W\n2zL7rHEczKk1TT4XJOa5AYDlTKqbl+MVWNIZia9Z+4GfPibGhUhM0lgcvyqnZ3SC\nmbc8jNkjzQMVXPW5RNlL0v0sL9HmdXn2l9aiXtWXFQKBgQD9/qbcBVFm3M9ZhI1U\nM2OvgrA+4OIOAAPzk37q2EwEeD2Ml3qLhphCGDx5M8XdjoviWdCdmOBkbSgrfbrJ\nAwPkm3bacgoY8mOF1LgWqDrrskhj7AzIplAelZi1Tl+jZgb500z8z9tw6SAPpE0Y\n9M+jaqpjuJL0ShSSBEPoy5ZmDwKBgQDv/DUhu+lNYEPXIpFckpfbST8Ltvs1D8xn\nwEFi8IBD6gWSsVMl2eBlvubrVfRPVOz4Q5oP1WZhS66jBsT1oXULM4ZP/IM8pC93\nnTEyELGngX5AkZ+XTNFg6EBURqkBLwUoSysC6g1QrxGoOd8MjR07oXHC04H4kKvL\nmyPlIYG/xwKBgQDjmxcbgwh3DNgid83N7l8eIBafwhXQReNEmp/ZAwrb1oQRQIzo\nCyzVzbIjSmm9KMfhhmqTTzOFyut8TabcCn+qlEXlatvxjwPwIbb1Vru2CoUkRZsP\nEzoorhWyIxFA82S86sTKp9kVTVMVsT4FZxvKbQNTaUsG7nuQfZH9QrL7sQKBgGnb\nKeSTarmCuOnwiYX385v6lyCnue8R5Ovu1ei6hRlv5L1sEm0jyeOXbWxpIbbMkyp/\nRzmsu9g29fVbSw2+Bk7Gj3xAufLJyVZNtCGNU4+7guEgWvD2JXVFfPqvoGSxROkd\nHv/OFcl9L6Cwxj5XEIn5EKELgs0No3nQGOk8+mqJAoGANE1PBf6FYniCeAGCjLq3\nydxaAvNANXAQ1lXPs2bMhcjdwNGmclU3IcmXzakIWkJGZNhp8BJyOHuephYzWILb\nNbw5zopMYy54Jmcqvv6EUp1e/wCiae4SvxBT0/vTo8V9CXA5r98iP3QOy9Gz0fUS\nRZU9zJu8M3GQQCoAPw2ou1w=\n-----END PRIVATE KEY-----\n",
+  "client_email": "blinc-985@focused-service-291115.iam.gserviceaccount.com",
+  "client_id": "109950894958441698631",
+  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+  "token_uri": "https://oauth2.googleapis.com/token",
+  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/blinc-985%40focused-service-291115.iam.gserviceaccount.com"
+}
