+++ audio_processing/source_seperation/asr_testing/manual_annotation.md
@@ -1,0 +1,17 @@
+Audio File | Annotated Text
+--- | ---
+
+"test_chunk1chan1.wav" | "can you like do like different lud light i didn't configure it but we definitely could change it  like raising keyboard i kind a liked the rainbow
+this one was picking up alright but this one for some reason is not
+got you
+it is okay i set up for this configuring but that would be just like how would you run the 
+what's a"
+
+"468_chunk0_025_chunk0_byte" | "discussion lined up which is even more yeah yeah just gave it up honestly i dont know how much cuz the mics are on top right a"
+
+"468_chunk0_025_chunk1_byte" | "do they sell cases for these we do actually they did actually make some cases but it doesnt cover the lights that much gotcha which is something that actually like we are taking over i might just turn off some i just dont know why i think they have them on"
+
+
+Sourse Seperated Audio File | Annotated Text
+--- | ---
+2022_2.wav | "hello hello bla bla bla and I think it is picking it is up banana you wanna go to the gym later to workout yeah yeah blob after we go to the gym can we go to the target yes we can go to the target after the gym hmm that is pretty cool yeah do you want to go to a place in pool i am going to beat you in pool yes i am wait make your voice sound emotional emotional say something emotional i lost my banana oh no i cant go to the gym okay okay so"
