+++ audio_processing/source_seperation/asr_testing/results.md
@@ -1,0 +1,10 @@
+## Word Error Rate Results
+
+Audio File | ASR Word Error Rate 
+--- | ---
+test_chunk1chan1 | 0.7538461538461538
+468_chunk0_025_chunk0_byte | 0.7777777777777778
+468_chunk0_025_chunk1_byte | 0.47058823529411764
+seperated_2022_2 | 0.7524752475247525
+2022_2 | 0.6435643564356436
+
