--- audio_processing/source_seperation/requirements.txt
@@ -1,1 +1,0 @@
-python
